+++ deleteUnit.php
@@ -1,7 +1,8 @@
 <?php
 
 include('config/db_connect.php');
-if (isset($_GET['id'])) {
+session_start();
+if (isset($_GET['id']) and $_SESSION['level'] != 0) {
     $id = $_GET['id'];
     $sql = "DELETE FROM db_donvi where madv = '$id'";
     $res = mysqli_query($conn, $sql);
