+++ deleteUser.php
@@ -1,7 +1,8 @@
 <?php
 
 include('config/db_connect.php');
-if (isset($_GET['id'])) {
+session_start();
+if (!isset($_GET['id']) and $_SESSION['id'] == 2) {
     $id = $_GET['id'];
     $sql = "DELETE FROM users where userid = '$id'";
     $res = mysqli_query($conn, $sql);
@@ -10,4 +11,6 @@ if (isset($_GET['id'])) {
     } else {
         echo "Error";
     }
+} else {
+    header('Location: index.php');
 }
